requirements.txt
earthengine-api
geopandas
rasterio
rioxarray
xarray
shapely
pyproj
pydantic
loguru
google-cloud-storage
pyyaml
streamlit
